# chord-chat/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chord chat",
    "version": "0.0.1",
    "description": "This is the documentation for chord chat. In this documentation you can view the different routes involved in the craft of chord chat. Chord chat is built with microservices, thus you can view the routes categorized under different services."
  },
  "servers": [
    {
      "url": " http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
  "paths": {
    "/signup": {
      "post": {
        "tags": [
          "User-Service"
        ],
        "description": "User signup route",
        "requestBody": {
          "description": "This body contains the email, password as required and other information about the route.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Unique email address of the user",
                    "example": "sayeed123@gmail.com"
                  },
                  "username": {
                    "type": "string",
                    "description": "username",
                    "example": "syd"
                  },
                  "password": {
                    "type": "string",
                    "description": "User's password",
                    "example": "12345"
                  },
                  "firstname": {
                    "type": "string",
                    "description": "User's first name",
                    "example": "sayeed"
                  },
                  "lastname": {
                    "type": "string",
                    "description": "User's last name",
                    "example": "safvan"
                  },
                  "talent": {
                    "type": "string",
                    "description": "User's talent",
                    "example": "guitarist"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "message": "Otp verification requested",
                "otp": 123
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Error message indicating email address is already in use"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/verify-signup": {
      "post": {
        "tags": [
          "User-Service"
        ],
        "description": "Verify OTP for signup route",
        "requestBody": {
          "description": "This body contains the OTP to be verified to proceed with signup.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "number",
                    "description": "OTP to be entered",
                    "example": 123
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "username": "sydsfn",
                  "email": "sayeed123@gmail.com",
                  "password": "123456",
                  "fans": [],
                  "idols": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Invalid OTP error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/signin": {
      "post": {
        "tags": [
          "User-Service"
        ],
        "description": "User signin route",
        "requestBody": {
          "description": "This body contains the email, password as required for user to signin",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "email address of the user",
                    "example": "sayeed@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "User's password",
                    "example": "12345"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "User's email address",
                      "example": "sayeed@gmail.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Error message indicating login request failed"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Error message indicating account not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/get-user": {
      "get": {
        "tags": [
          "User-Service"
        ],
        "description": "Get single user route",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "User logged in with email and password",
                      "example": "sayeed@gmail.com"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Error message indicating login request failed"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/signout": {
      "get": {
        "tags": [
          "User-Service"
        ],
        "description": "User signout route",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "message": "User signed out successfully!"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "No user is currently logged in!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user-profile": {
      "get": {
        "tags": [
          "User-Service"
        ],
        "description": "User Profile details",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "gigs": [],
                    "fans": 0,
                    "idols": 0,
                    "profile_description": "enter description",
                    "username": "sydonguitars",
                    "pfp": "12312nfajisrajwrwor.jpg"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "No user is currently logged in!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/edit-profile": {
      "patch": {
        "tags": [
          "User-Service"
        ],
        "description": "User edit profile details",
        "requestBody": {
          "description": "This body contains the tweaked details of user profile",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "username of the user",
                    "example": "sydonguitarsupdated"
                  },
                  "oldpassword": {
                    "type": "string",
                    "description": "User's old password",
                    "example": "12345"
                  },
                  "newpassword": {
                    "type": "string",
                    "description": "User's new password",
                    "example": "123"
                  },
                  "confirmpassword": {
                    "type": "string",
                    "description": "Confirm password",
                    "example": "123"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "message": "Profile updated successfully"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "No user is currently logged in!"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Passwords doesnt match issue!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/find-users": {
      "post": {
        "tags": [
          "User-Service"
        ],
        "description": "Users fetching based on search results",
        "requestBody": {
          "description": "This body contains search term to search",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "searchTerm": {
                    "type": "string",
                    "description": "username of the user",
                    "example": "syd"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "data": [
                  {
                    "id": 2,
                    "username": "sydonguitars",
                    "email": "sayeed@gmail.com",
                    "password": "12345"
                  },
                  {
                    "id": 1,
                    "username": "sydsfn",
                    "email": "sayeed123@gmail.com",
                    "password": "123456"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "No user is currently logged in!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "No users found!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/forgot-password": {
      "post": {
        "tags": [
          "User-Service"
        ],
        "description": "Forgot password submission route",
        "requestBody": {
          "description": "This body contains email and will produce otp for futher reset proceedings",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "email to reset",
                    "example": "sayeed123@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "message": "Otp verification required",
                  "otp": 123
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Invalid OTP error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/verify-forgot-password": {
      "post": {
        "tags": [
          "User-Service"
        ],
        "description": "Verify and save new password for the requested user",
        "requestBody": {
          "description": "This body contains the OTP to be verified, new password and confirm password fields, to proceed with resetting the password.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "number",
                    "description": "OTP to be entered",
                    "example": 123
                  },
                  "password": {
                    "type": "string",
                    "description": "New password",
                    "example": "12345"
                  },
                  "confirmpassword": {
                    "type": "string",
                    "description": "Confirm password",
                    "example": "12345"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "username": "sydsfn",
                  "email": "sayeed123@gmail.com",
                  "password": "123456",
                  "fans": [],
                  "idols": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Invalid OTP error"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Passwords doesnt match"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/admin-get-all-users": {
      "get": {
        "tags": [
          "User-Service"
        ],
        "description": "Route to get all the users!",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "username": "sydsfn",
                    "email": "sayeed123@gmail.com",
                    "password": "123456",
                    "fans": [],
                    "idols": []
                  },
                  {
                    "id": 2,
                    "username": "sydsf2",
                    "email": "sayeed122@gmail.com",
                    "password": "123453",
                    "fans": [],
                    "idols": []
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad request error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Bad request error"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Invalid OTP error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}